refactor: rename credential exchange record file
