+++ packages/core/src/modules/credentials/protocol/v2/handlers/V2OfferCredentialHandler.ts
@@ -7,7 +7,7 @@ import type { MediationRecipientService } from '../../../../routing/services/Med
 import type { CredentialFormatService } from '../../../formats/CredentialFormatService'
 import type { HandlerAutoAcceptOptions } from '../../../formats/models/CredentialFormatServiceOptions'
 import type { CredentialPreviewAttribute } from '../../../models/CredentialPreviewAttributes'
-import type { CredentialExchangeRecord } from '../../../repository/CredentialRecord'
+import type { CredentialExchangeRecord } from '../../../repository/CredentialExchangeRecord'
 import type { V2CredentialService } from '../V2CredentialService'
 
 import { AriesFrameworkError } from '../../../../../../src/error/AriesFrameworkError'
