+++ packages/core/src/modules/credentials/repository/CredentialRepository.ts
@@ -4,7 +4,7 @@ import { InjectionSymbols } from '../../../constants'
 import { Repository } from '../../../storage/Repository'
 import { StorageService } from '../../../storage/StorageService'
 
-import { CredentialExchangeRecord } from './CredentialRecord'
+import { CredentialExchangeRecord } from './CredentialExchangeRecord'
 
 @scoped(Lifecycle.ContainerScoped)
 export class CredentialRepository extends Repository<CredentialExchangeRecord> {
