+++ packages/core/src/modules/credentials/repository/index.ts
@@ -1,3 +1,3 @@
-export * from './CredentialRecord'
+export * from './CredentialExchangeRecord'
 export * from './CredentialRepository'
 export * from './CredentialMetadataTypes'
